Revert "[NFC] Fix CodeQL violations in compiler-rt. (#157793)"
This reverts commit b44e6e01f7f778bbb569e07821f5ddfdb90c4d4d.

+++ compiler-rt/lib/fuzzer/FuzzerCorpus.h
@@ -336,8 +336,7 @@ public:
   void PrintFeatureSet() {
     for (size_t i = 0; i < kFeatureSetSize; i++) {
       if(size_t Sz = GetFeature(i))
-        Printf("[%zd: id %zd sz%zd] ", i, (size_t)SmallestElementPerFeature[i],
-               Sz);
+        Printf("[%zd: id %zd sz%zd] ", i, SmallestElementPerFeature[i], Sz);
     }
     Printf("\n\t");
     for (size_t i = 0; i < Inputs.size(); i++)
